deployment error fixed

+++ app/api/conversations/[conversationId]/route.ts
@@ -7,7 +7,7 @@ interface IParams {
   conversationId?: string;
 }
 
-export default async function DELETE(
+export default async function handler(
   req: Request,
   {
     params,
@@ -15,49 +15,53 @@ export default async function DELETE(
     params: IParams;
   }
 ) {
-  try {
-    const { conversationId } = params;
-    const currentUser = await getCurrentUser();
+  if (req.method === 'DELETE') {
+    try {
+      const { conversationId } = params;
+      const currentUser = await getCurrentUser();
 
-    if (!currentUser?.id || !currentUser.email) {
-      return new NextResponse("Unauthorized", { status: 401 });
-    }
-
-    if (!conversationId) {
-      return new NextResponse("Bad Request", { status: 400 });
-    }
-
-    const conversation = await prisma.conversation.findUnique({
-      where: {
-        id: conversationId,
-      },
-      include: {
-        users: true,
-      },
-    });
-
-    if (!conversation) {
-      return new NextResponse("Not Found", { status: 404 });
-    }
-
-    const deletedConversation = await prisma.conversation.deleteMany({
-      where: {
-        id: conversationId,
-        userIds: {
-          hasSome: [currentUser.id],
-        },
-      },
-    });
-
-    conversation.users.forEach((user) => {
-      if (user.email) {
-        pusherServer.trigger(user.email, "conversation:remove", conversation);
+      if (!currentUser?.id || !currentUser.email) {
+        return new NextResponse("Unauthorized", { status: 401 });
       }
-    });
 
-    return NextResponse.json(deletedConversation);
-  } catch (error) {
-    console.log("[CONVERSATION_DELETE_ERROR]", error);
-    return new NextResponse("Internal Server Error", { status: 500 });
+      if (!conversationId) {
+        return new NextResponse("Bad Request", { status: 400 });
+      }
+
+      const conversation = await prisma.conversation.findUnique({
+        where: {
+          id: conversationId,
+        },
+        include: {
+          users: true,
+        },
+      });
+
+      if (!conversation) {
+        return new NextResponse("Not Found", { status: 404 });
+      }
+
+      const deletedConversation = await prisma.conversation.deleteMany({
+        where: {
+          id: conversationId,
+          userIds: {
+            hasSome: [currentUser.id],
+          },
+        },
+      });
+
+      conversation.users.forEach((user) => {
+        if (user.email) {
+          pusherServer.trigger(user.email, "conversation:remove", conversation);
+        }
+      });
+
+      return NextResponse.json(deletedConversation);
+    } catch (error) {
+      console.log("[CONVERSATION_DELETE_ERROR]", error);
+      return new NextResponse("Internal Server Error", { status: 500 });
+    }
+  } else {
+    return new NextResponse("Method Not Allowed", { status: 405 });
   }
 }
